Changes to stac_stocks
Alright, here this goes.
The developer of StAC is working on making the anticheat invisible... But what's the point if cheaters can find out that it's running on the server just by getting banned or seeing someone else get banned? They can just come back and bypass the anticheat on an alt account if that happens.
So, ban messages should be as simple as possible ("Banned from server"), and all other players on the server should just see that as well. If they have questions or want to know why, just tell them that the player who was banned was cheating.

The only people who should have any idea of what is going on with the anticheat are any admins, moderators, or anyone who can be trusted. Regular players should not ever know that the anticheat is running on the server, or at the very least, shouldn't know what anticheat is running on the server or what detections there are.