Create sbpstac.sp
This is my fork of sbp. It is incompatible with the regular sbp, but also required in order to prevent the plugins from showing up in the plugin list. I guarantee a better solution than this exists, but this worked for me. Won't work for all servers since some might actually use the plugins we impersonate, creating duplicate entries.